Expand native settings controls and clarify save feedback
