+++ docs/MOD_SETTINGS_CONTROLS.md
@@ -1,21 +1,26 @@
-# First mod settings controls
+# Mod settings controls
 
 Build real controls on the navigation foundation in small slices. Register only
 working controls; omit empty groups. Stable setting keys and storage owners stay
 independent of labels and placement.
 
-## Initial layout
+See [the current architecture contract](MOD_SETTINGS.md) and
+[native adapter ownership](MOD_SETTINGS_NATIVE_ADAPTER.md).
+
+## Layout
 
 | Location | Control | Existing owner |
 | --- | --- | --- |
-| Mod Settings > User Interface | Instant warp mode: Normal (ask), Warp, Jump | `ui.auto_confirm_instant_warp` and the Alt+I action |
-| Mod Settings > Graphics > Fleet Labels | Player label detail and zoom threshold | `graphics.zoom_label_player_detail`, `graphics.zoom_label_player_threshold` |
-| Mod Settings > Graphics > Fleet Labels | Non-player label detail and zoom threshold | `graphics.zoom_label_non_player_detail`, `graphics.zoom_label_non_player_threshold` |
+| Mod Settings > Map & Travel | Instant warp mode: Normal (ask), Warp, Jump | `ui.auto_confirm_instant_warp` and the Alt+I action |
+| Mod Settings > Fleet Labels | Player label detail and zoom threshold | `graphics.zoom_label_player_detail`, `graphics.zoom_label_player_threshold` |
+| Mod Settings > Fleet Labels | Non-player label detail and zoom threshold | `graphics.zoom_label_non_player_detail`, `graphics.zoom_label_non_player_threshold` |
+| Mod Settings > Camera | Keyboard zoom speed and pan glide | `graphics.keyboard_zoom_speed`, `graphics.system_pan_momentum_falloff` |
+| Mod Settings > Previews & Cargo | Locate/Recall while previewing; automatic cargo and target types | Existing preview/cargo keys in `[ui]` |
 | Future separate branch: Hotkeys | Rebind existing actions | Existing shortcut parser and `MapKey` registrations |
 | General > confirmation page | Confirm Forbidden Tech upgrades | Inverse of `ui.auto_confirm_ft_upgrade` |
 
-The controls branch implements instant warp, Fleet Labels and Forbidden Tech on
-Windows x64. Hotkey editing remains a separate branch. Native confirmation
+The controls branch implements these controls on Windows x64.
+Hotkey editing remains a separate branch. Native confirmation
 controls stay on the native page. FC retains its existing owner.
 
 ## Instant warp mode
@@ -28,7 +33,8 @@ the picker changes only the global fallback mode.
 
 Reuse the existing single runtime writer, optimistic conflict handling and
 source-preserving TOML edits. UI readback confirms the live value, not durable
-storage; asynchronous save failures continue to go to the log. Reopening must
+storage; asynchronous failures produce a quiet session-only notice, with details
+in the log. Reopening must
 read the current owner, and shortcut changes must refresh a visible selector.
 Native selection callbacks need the same rendering, stale-context and reentry
 protection already exercised for boolean controls.
@@ -118,18 +124,25 @@ or interior entries using unwind metadata. Client SHA256:
 `487af4bb9c697c353be9714359a97dddcece5dab872622a6c498a27bbfc44f40`.
 This is Windows evidence, not proof of macOS hook fit or native widget behavior.
 
+## Camera and previews
+
+Keyboard zoom speed offers 0–1000 in steps of 25 with whole-number labels. Pan
+glide offers 0–0.99 in steps of 0.01; it retains the existing pan formula. These
+are UI editing ranges, not new TOML constraints. Out-of-range loaded values are
+preserved and explained rather than clamped. The shared slider path controls
+display precision without writing a loaded value.
+
+Preview toggles and their existing hotkeys call the same owner, so live state,
+readback and saving agree. Locate/Recall labels invert their stored disable
+flags. Cargo targets are visible only while auto-open is ON, with their saved
+preferences retained while hidden. Hook installation success gates each group.
+
 ## Future organization and commands (design notes)
 
-Use the existing TOML sections as the organizing vocabulary: Audio (`[audio]`),
-Buffs (`[buffs]`), Config (`[config]`), Control (`[control]`), Graphics, and so on.
-Introduce a group when it gains a working control and an explicit apply path.
-Do not populate empty groups or build a generic editor for every config key.
-Human labels and nested pages can be clearer than raw keys; changing placement
-must not change storage identity. The current populated groups are Graphics
-(`[graphics]`, Fleet Labels) and User Interface (`[ui]`, Instant warp mode).
-Control (`[control]`) can be introduced with its own working controls, such as
-hotkeys; instant warp is not moved into that TOML section. Confirmations continue
-to use the native confirmation page.
+Use player tasks for navigation and TOML sections as storage references.
+Introduce a group only when it gains a working control and explicit apply path.
+Changing placement must not change storage identity. Confirmations continue on
+the native confirmation page; arbitrary TOML keys are not discovered as controls.
 
 A future **Restart client** command could support controls that explicitly need
 restart. It would perform an ordinary client restart, settle pending saves using
