+++ mods/src/config.h
@@ -268,7 +268,7 @@ public:
   bool installDailyFactionBulkClaimHooks;
   bool installInstantWarpConfirmationHooks;
   bool installAudioEventHooks;
-  bool installModConfirmationSettings;
+  bool installNativeSettings;
 
   std::string config_settings_url;
   std::string config_assets_url_override;
