+++ mods/src/patches/parts/runtime_config.cc
@@ -188,7 +188,16 @@ void Configure(const toml::table& loaded)
                                        {"graphics", "zoom_label_non_player_detail"},
                                        {"graphics", "zoom_label_player_threshold"},
                                        {"graphics", "zoom_label_non_player_threshold"},
-                                       {"ui", "auto_confirm_ft_upgrade"}}) {
+                                       {"graphics", "keyboard_zoom_speed"},
+                                       {"graphics", "system_pan_momentum_falloff"},
+                                       {"ui", "auto_confirm_ft_upgrade"},
+                                       {"ui", "disable_preview_locate"},
+                                       {"ui", "disable_preview_recall"},
+                                       {"ui", "show_cargo_default"},
+                                       {"ui", "show_player_cargo"},
+                                       {"ui", "show_station_cargo"},
+                                       {"ui", "show_hostile_cargo"},
+                                       {"ui", "show_armada_cargo"}}) {
       std::optional<config_edit::Value> value;
       auto                              node = loaded[section][key];
       if (node.is_boolean())
