+++ mods/src/patches/patches.cc
@@ -48,7 +48,7 @@ void InstallDoubleClickAssignShipHooks();
 void InstallInstantWarpConfirmationHooks();
 void InstallForbiddenTechConfirmationHooks();
 void InstallAudioEventHooks();
-void InstallModConfirmationSettings();
+void InstallNativeSettings();
 
 __int64 il2cpp_init_hook(auto original, const char* domain_name)
 {
@@ -123,7 +123,7 @@ __int64 il2cpp_init_hook(auto original, const char* domain_name)
   spdlog::info("Initializing code hooks:");
   bool install_forbidden_tech = cfg.auto_confirm_ft_upgrade;
 #if defined(_WIN32) && defined(_M_X64)
-  install_forbidden_tech |= cfg.installModConfirmationSettings;
+  install_forbidden_tech |= cfg.installNativeSettings;
 #endif
   const PatchEntry patches[] = {
       {"UiScaleHooks", {InstallUiScaleHooks, &cfg.installUiScaleHooks}},
@@ -155,7 +155,8 @@ __int64 il2cpp_init_hook(auto original, const char* domain_name)
       {"InstantWarpConfirm", {InstallInstantWarpConfirmationHooks, &cfg.installInstantWarpConfirmationHooks}},
       {"ForbiddenTechConfirm", {InstallForbiddenTechConfirmationHooks, &install_forbidden_tech}},
       {"AudioEvents", {InstallAudioEventHooks, &cfg.installAudioEventHooks}},
-      {"ModConfirmationSettings", {InstallModConfirmationSettings, &cfg.installModConfirmationSettings}},
+      // Retain the existing debug patch key; this installer owns both settings surfaces.
+      {"ModConfirmationSettings", {InstallNativeSettings, &cfg.installNativeSettings}},
   };
   printf("il2cpp_init_hook(%s)\n", domain_name);
