+++ mods/src/settings/camera_settings.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "slider_setting.h"
+
+namespace mod_settings
+{
+SliderSetting& KeyboardZoomSpeedSetting();
+SliderSetting& PanGlideSetting();
+bool           KeyboardZoomControlAvailable();
+bool           PanGlideControlAvailable();
+} // namespace mod_settings
