+++ mods/src/settings/fleet_labels.h
@@ -5,6 +5,7 @@ namespace mod_settings
 {
 ChoiceSetting& FleetLabelDetailSetting(bool player);
 SliderSetting& FleetLabelThresholdSetting(bool player);
+std::string    FleetLabelSummary(bool player);
 } // namespace mod_settings
 // Existing zoom adapter owns native fleet labels and reports installation.
 bool FleetLabelControlsAvailable();
