+++ mods/src/settings/native/action_widgets.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#if defined(_WIN32) && defined(_M_X64)
+#include "interop.h"
+#include "settings/page_catalog.h"
+
+namespace mod_settings::native
+{
+using ActionRow = std::pair<ActionSetting*, std::size_t>;
+bool      ActionsActive();
+void      RefreshActions();
+ActionRow ActionFor(Il2CppObject* context);
+void      AddActionRow(Il2CppObject* director, Il2CppObject* context, Il2CppObject* parent, ActionSetting& action,
+                       std::size_t index);
+void      SyncActionRows(Il2CppObject* controller, Il2CppObject* context, const PageCatalog::Page& page);
+void      InstallActionWidgets();
+} // namespace mod_settings::native
+
+#endif
