+++ tests/page_catalog_test.cc
@@ -80,6 +80,53 @@ int main()
   SliderSetting otherSlider({"other.zoom", "Threshold", [] { return ValueReadResult<float>::Known(0.5f, 1); },
                              [](float, std::uint64_t) { return ApplyResult::Applied; }},
                             0, 1, 0.01f, [] { return true; });
+  // Command registration/building is presentation-only: no invocation or read.
+  int           commandCalls = 0;
+  std::size_t   commandCount = 2;
+  ActionSetting command{"record", "Record",
+                        [&](std::size_t) {
+                          ++commandCalls;
+                          return ActionSetting::Presentation{};
+                        },
+                        [&](std::size_t) { ++commandCalls; },
+                        [&] {
+                          ++commandCalls;
+                          return commandCount;
+                        }};
+  PageCatalog   commands("commands", "Commands");
+  assert(commands.AddAction("commands", command) == Registration::Added);
+  assert(commands.AddAction("commands", command) == Registration::Duplicate);
+  const auto commandPlan = commands.Build();
+  assert(commandPlan.size() == 1 && commandPlan.front().ControlRows() == 1 && commandCalls == 0);
+  assert(commands.AddAction("commands", command) == Registration::Frozen);
+  // Registration counts definitions without evaluating dynamic rows. Repeated
+  // native identities survive shrink/regrowth and reject noncanonical suffixes.
+  assert(command.item_index(command.item_id(42)) == 42);
+  for (const auto* id : {"record", "record.row.", "record.row.-1", "record.row.01", "record.row.1x", "other.row.1",
+                         "record.row.999999999999999999999999999999999999"})
+    assert(!command.item_index(id));
+  command.read = [&](std::size_t index) {
+    ++commandCalls;
+    return ActionSetting::Presentation{std::to_string(index), "Change", "", true, true};
+  };
+  assert(command.Read(1).actionable());
+  commandCount          = 1;
+  const auto beforeRead = commandCalls;
+  assert(!command.Read(1).visible && commandCalls == beforeRead + 1); // Count only; no stale reader call.
+  assert(!command.Read(1).actionable());
+  commandCount = 2;
+  assert(command.Read(1).label == "1");
+  auto presentation    = command.Read(0);
+  presentation.visible = false;
+  assert(!presentation.actionable());
+  presentation.visible = true;
+  presentation.button.clear();
+  assert(!presentation.actionable()); // Information rows do not expose a blank button.
+  presentation.button  = "Change";
+  presentation.enabled = false;
+  assert(!presentation.actionable());
+  assert(commandPlan[0].PositionFor(command.id()) == 0);
+  assert(commandPlan[0].PositionFor("unknown") == commandPlan[0].items.size());
   PageCatalog   combined("labels", "Fleet Labels");
   assert(combined.AddHeading("labels", "player.heading", "Player", true) == Registration::Added);
   assert(combined.AddChoice("labels", player) == Registration::Added);
@@ -113,6 +160,40 @@ int main()
   assert(!bounded[0].SectionFor("player")); // Controls before a heading are unaffected.
   assert(bounded[0].SectionFor("player.zoom")->id == "collapsible");
   assert(!bounded[0].SectionFor("other.zoom")); // A plain heading ends a collapsible section.
+  bool           masterOn = false, masterAvailable = true;
+  int            visibilityReads = 0;
+  BooleanSetting master({"master",
+                         "Master",
+                         [&] {
+                           ++visibilityReads;
+                           return masterAvailable ? ReadResult::Known(masterOn, 1) : ReadResult{};
+                         },
+                         {}});
+  PageCatalog    conditional("conditional", "Conditional sections");
+  assert(conditional.AddBoolean("conditional", master) == Registration::Added);
+  assert(conditional.AddHeading("conditional", "targets", "Targets", false, [&] {
+    const auto state = master.Observe().state;
+    return state.known() && *state.value;
+  }) == Registration::Added);
+  assert(conditional.AddBoolean("conditional", setting) == Registration::Added);
+  assert(conditional.AddHeading("conditional", "later", "Later section") == Registration::Added);
+  assert(conditional.AddSlider("conditional", playerSlider) == Registration::Added);
+  const auto  conditionalPlan = conditional.Build();
+  const auto& page            = conditionalPlan.front();
+  assert(visibilityReads == 0); // Building retains hidden controls without reading their dependency.
+  assert(page.HasConditionalSections() && !bounded[0].HasConditionalSections());
+  const auto writesBeforeVisibility = writes;
+  for (bool enabled : {false, true, false, true}) {
+    masterOn = enabled; // A live change from either UI or shortcut uses the same reader.
+    assert(page.IsVisible("master"));
+    assert(page.IsVisible("targets") == enabled);
+    assert(page.IsVisible(setting.id()) == enabled);
+    assert(page.IsVisible("later") && page.IsVisible(playerSlider.state().id()));
+    assert(value && writes == writesBeforeVisibility); // Hiding/revealing never clears a target preference.
+  }
+  masterAvailable = false;
+  assert(!page.IsVisible("targets") && !page.IsVisible(setting.id()));
+  assert(page.IsVisible("master") && page.IsVisible("unknown"));
   bool        rejected = false;
   std::thread wrong_thread([&] {
     try {
