+++ tests/slider_setting_test.cc
@@ -23,11 +23,12 @@ int main()
                            value = desired;
                            return ApplyResult::Applied;
                          }},
-                        0, 1, 0.01f, [&] { return enabled; });
+                        0, 1, 0.01f, [&] { return enabled; }, SliderLabel::Percentage, 2, "Enable test feature");
   NativeViewState view(setting), stale(setting);
   view.Bind();
   stale.Bind();
   assert(view.known() && view.number() == 0.33333f); // Reading does not quantize player-authored values.
+  assert(view.displayNumber() == 0.33f && value == 0.33333f && writes == 0);
   assert(view.Request(0.604f) == Outcome::AppliedVerified && std::abs(value - 0.6f) < 0.000001f);
   assert(stale.Request(0.8f) == Outcome::Conflict && writes == 1);
   assert(view.Request(std::numeric_limits<float>::quiet_NaN()) == Outcome::Rejected);
@@ -35,6 +36,7 @@ int main()
   assert(view.Request(-0.01f) == Outcome::Rejected && view.Request(1.01f) == Outcome::Rejected);
   enabled = false;
   assert(view.known() && !view.enabled() && view.Request(0.9f) == Outcome::Rejected && writes == 1);
+  assert(view.disabledReason() == "Enable test feature");
   enabled = true;
   {
     NativeViewState::RenderScope render(view);
@@ -49,6 +51,7 @@ int main()
   value     = std::numeric_limits<float>::quiet_NaN();
   view.Bind();
   assert(!view.known());
+  assert(view.unavailableReason() == "Invalid value; edit TOML");
   value = 0.5f;
   view.Bind();
   view.Unbind();
@@ -59,6 +62,32 @@ int main()
   assert(catalog.AddSlider("labels", setting) == Registration::Duplicate);
   assert(catalog.Build().size() == 2);
   assert(catalog.AddSlider("labels", setting) == Registration::Frozen);
+
+  // Native drag labels can arrive after a change has already snapped/read back.
+  // Display that snapshot, and keep display precision independent of the step.
+  value = 382.1429f;
+  SliderSetting speed(
+      {"speed", "Speed", [&] { return ValueReadResult<float>::Known(value, 1); },
+       [&](float desired, std::uint64_t) {
+         value = desired;
+         return ApplyResult::Applied;
+       }},
+      0, 1000, 25, [] { return true; }, SliderLabel::Value, 0);
+  NativeViewState speedView(speed);
+  speedView.Bind();
+  assert(speedView.disabledReason().empty()); // Generic sliders have no feature-specific instruction.
+  value = 2000.0f;
+  speedView.Bind();
+  assert(!speedView.known() && speedView.unavailableReason() == "Out of range; edit TOML");
+  speedView.Bind(); // Reopening cannot repair a value outside this editor's range.
+  assert(value == 2000.0f && speedView.unavailableReason() == "Out of range; edit TOML");
+  value = 382.1429f;
+  speedView.Bind();
+  assert(speedView.displayNumber() == 382 && value == 382.1429f);
+  assert(speedView.Request(382.1429f) == Outcome::AppliedVerified);
+  assert(speedView.number() == 375 && speedView.displayNumber() == 375);
+  assert(setting.DisplayValue(0.72575f) == 0.73f);
+  assert(setting.DisplayValue(0.0f) == 0 && setting.DisplayValue(0.99f) == 0.99f);
 
   Il2CppType single{}, nothing{}, integer{};
   single.type  = IL2CPP_TYPE_R4;
